Client and Server packets transition to using metaclasses.
This is a very complicated patch to simplify the work required to define server
and client packets. When a class CP_FOO(CP) is defined, it is automatically
put into a lookup table, and also instanciated and put into the namespace as
cp_foo.

Further work: put all the server and client packets into their own module, 
and don't do automatic name space polution.