process all pending network data
On a test server with planets moving, and a low update rate, the planet
move packets were delayed by one update time before being displayed on
screen.  This was caused by data waiting in the TCP socket that was
ignored.

The fix is to process all data in both the UDP and TCP sockets before
returning.