short term fixes for microsoft windows
According to a contributor, on Microsoft Windows the server list screen
does not respond, this was found to be because the event queue wasn't
being serviced fast enough.  This patch reduces the user event demand.

Also, MSG_DONTWAIT is not defined.  This patch removes that usage if it
is not present.