handle all display events in wait mode
On Microsoft Windows, according to a contributor, on the server list
screen, the high rate of user events causes the pygame event queue to
fill, so we pull all events from the queue before waiting.  This may be
due to the UDP receive call blocking or taking a long time due to
firewall software.